macOS: Ensure initial geometry is always set
QPlatformWindow initializes its view of the geometry based on the
QWindow geometry during construction. If the initial geometry we
then compute is the same, we would end up exiting early from
QCocoaWindow::setGeometry(), because we compared the new geometry
against the QPlatformWindow::geometry(), and the geometry would
never be reflected in the NSView.

Due to other setGeometry calls this was in most cases masked, but
could in theory happen, and is preventing us from cleaning up other
parts of the code.

The call to QWindow::setGeometry() after setting the initial geometry
is also broken, as the initial geometry is available through the
platform window and QWindow::geometry() already, so setting it again
serves nothing except disabling d->positionAutomatic, which is not
correct.

Change-Id: I0db3cfe7d7c3d14070beee6ae3ea3dfd49da9e05
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>